requirements.txt
appnope==0.1.0
astroid==2.2.5
backcall==0.1.0
decorator==4.4.0
Django==2.2.4
django-extensions==2.2.1
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
django-ckeditor
django-import-export==2.4.0